// src/types/api/product.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * App API
 * OpenAPI spec version: 1.0.0
 */
import type { ProductCategory } from './productCategory';
import type { Review } from './review';

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  /** 価格（円） */
  price: number;
  inStock: boolean;
  description: string;
  /**
     * @minimum 0
     * @maximum 5
     */
  rating: number;
  reviews: Review[];
}